# Conflicts:
#	src/main/java/com/example/backend/controllers/AdminController.java
#	src/main/java/com/example/backend/controllers/AppController.java
#	src/main/java/com/example/backend/controllers/CardController.java
#	src/main/java/com/example/backend/controllers/PurchaseController.java
#	src/main/java/com/example/backend/model/data/App.java
#	src/main/java/com/example/backend/model/data/Review.java
#	src/main/java/com/example/backend/model/data/Subscription.java
#	src/main/java/com/example/backend/repositories/AppRepository.java
#	src/main/java/com/example/backend/repositories/CardRepository.java
#	src/main/java/com/example/backend/repositories/ReviewRepository.java
#	src/main/java/com/example/backend/repositories/SubscriptionRepository.java
#	src/main/java/com/example/backend/repositories/data/finances/PurchaseRepository.java
#	src/main/java/com/example/backend/services/auth/RoleManagementService.java
#	src/main/java/com/example/backend/services/data/AppService.java
#	src/main/java/com/example/backend/services/data/BudgetService.java
#	src/main/java/com/example/backend/services/data/CardService.java
#	src/main/java/com/example/backend/services/data/PurchaseService.java
#	src/main/java/com/example/backend/services/data/ReviewService.java

+++ src/main/java/com/example/backend/dto/data/budget/BudgetStatusDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class BudgetStatusDto {
-    private Float budget;
-    private Float spending;
-    private Float remaining;
+    private Double budget;
+    private Double spending;
+    private Double remaining;
 }
